Squashed commit of the following:
commit 6293066423136a50888bc4619ed32bec3960075a
Author: andrew-guz <andrew.guz.spb@gmail.com>
Date:   Tue Feb 7 17:19:15 2023 +0300

    Error fixed.

commit d194b71b3fd193818db75d5dd2cdb68dc79026e8
Author: andrew-guz <andrew.guz.spb@gmail.com>
Date:   Tue Feb 7 17:13:51 2023 +0300

    Error fixed.

commit 374511f2cd518dec63f3a5004a672812e29f1ccd
Author: andrew-guz <andrew.guz.spb@gmail.com>
Date:   Tue Feb 7 17:07:12 2023 +0300

    Use one client and do this faster. Configurable verbose. Additional time for measurements.

commit 02c14cca69a977998ec874ac4002560f0cf60bba
Author: andrew-guz <andrew.guz.spb@gmail.com>
Date:   Tue Feb 7 16:26:47 2023 +0300

    Arduino over https

commit 0ff242cbfaa7a2d9ace2b38a056317b4757040a6
Author: andrew-guz <andrew.guz.spb@gmail.com>
Date:   Tue Feb 7 15:52:09 2023 +0300

    Not Thermometer simulator and FE to BE also uses https.

commit 3c0fc00d5de67c3247514231480528e5bda03ba2
Author: andrew-guz <andrew.guz.spb@gmail.com>
Date:   Tue Feb 7 15:15:39 2023 +0300

    Start BE service with https.

commit c6ecacff56a29ff0c07c9c1a0b0925b9b24ff302
Author: Andrew Guz <103986501+andrew-guz@users.noreply.github.com>
Date:   Tue Feb 7 13:49:45 2023 +0300

    Update HOW_TO_RUN.md

commit 63849572d43e7668d79b1dc82054188a4bf84b48
Author: andrew-guz <andrew.guz.spb@gmail.com>
Date:   Tue Feb 7 13:48:24 2023 +0300

    Possibility to run Frontend service in https mode.

+++ Cpp/UniversalDeviceFrontendService/source/stackWidgets/TemperatureChartModel.cpp
@@ -35,7 +35,7 @@ int TemperatureChartModel::rowCount(const WModelIndex& parent) const
 WModelIndex TemperatureChartModel::index(int row, int column, const WModelIndex& parent) const
 {
     if (row >= 0 &&
-        row < _data.size() &&
+        (size_t)row < _data.size() &&
         column >= 0 &&
         column < 2 &&
         !parent.isValid())
@@ -52,7 +52,7 @@ cpp17::any TemperatureChartModel::data(const WModelIndex& index, ItemDataRole ro
 {
     if (index.isValid() &&
         index.row()>= 0 &&
-        index.row() < _data.size() &&
+        (size_t)index.row() < _data.size() &&
         index.column() >= 0 &&
         index.column() < 2  &&
         role == ItemDataRole::Display)
